[KS] BIG update!! VS Menu translated
This is one of the biggest updates of the translation itself, other than the Single Player Menu (almost done)!  VS Menu is about 80%-ish complete, with one minor graphical glitch, a few spacing issues, and minor text to translate using different fonts.  As I have been working on other parts of the game, I didn't want everyone to have to wait for it to be perfect before a first look.  Enjoy!